+++ packages/managers/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @module-federation/managers
 
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies [80af3f3]
+  - @module-federation/sdk@0.1.18
+
+## 0.1.17
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.17
+
+## 0.1.16
+
+### Patch Changes
+
+- cce5404: chore: support normalize shared.import option
+- Updated dependencies [364f2bc]
+  - @module-federation/sdk@0.1.16
+
 ## 0.1.15
 
 ### Patch Changes