+++ packages/runtime-tools/CHANGELOG.md
@@ -1,5 +1,30 @@
 ## [1.0.1-canary.1](https://github.com/module-federation/core/compare/runtime-1.0.0...runtime-1.0.1-canary.1) (2023-12-06)
 
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies [80af3f3]
+  - @module-federation/runtime@0.1.18
+  - @module-federation/webpack-bundler-runtime@0.1.18
+
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [26bff6e]
+  - @module-federation/runtime@0.1.17
+  - @module-federation/webpack-bundler-runtime@0.1.17
+
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [103cd07]
+- Updated dependencies [425fc9d]
+  - @module-federation/runtime@0.1.16
+  - @module-federation/webpack-bundler-runtime@0.1.16
+
 ## 0.1.15
 
 ### Patch Changes