+++ packages/third-party-dts-extractor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @module-federation/third-party-dts-extractor
 
+## 0.1.18
+
+## 0.1.17
+
+## 0.1.16
+
 ## 0.1.15
 
 ## 0.1.14